+++ backend/contracts/RaffleProxy.sol
@@ -1,0 +1,132 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.28;
+
+contract RaffleProxy {
+    bytes32 private constant IMPLEMENTATION_SLOT =
+        bytes32(uint256(keccak256("eip1967.proxy.implementation")) - 1);
+
+    bytes32 private constant ADMIN_SLOT =
+        bytes32(uint256(keccak256("eip1967.proxy.admin")) - 1);
+
+    constructor() {
+        _initialize();
+    }
+
+    modifier ifAdmin() {
+        if (msg.sender == _getAdmin()) {
+            _;
+        } else {
+            _fallback();
+        }
+    }
+
+    function _initialize() internal {
+        _setAdmin(msg.sender);
+    }
+
+    function changeAdmin(address _admin) external ifAdmin {
+        if (_admin == _getAdmin()) {
+            revert();
+        }
+
+        _setAdmin(_admin);
+    }
+
+    function _delegateCall(address _implementation) private {
+        assembly {
+            // Copy msg.data. We take full control of memory in this inline assembly
+            // block because it will not return to Solidity code. We overwrite the
+            // Solidity scratch pad at memory position 0.
+            calldatacopy(0x00, 0x00, calldatasize())
+
+            // Call the implementation.
+            // out and outsize are 0 because we don't know the size yet.
+            let result := delegatecall(
+                gas(),
+                _implementation,
+                0x00,
+                calldatasize(),
+                0x00,
+                0x00
+            )
+
+            // Copy the returned data.
+            returndatacopy(0x00, 0x00, returndatasize())
+
+            switch result
+            // delegatecall returns 0 on error.
+            case 0 {
+                revert(0x00, returndatasize())
+            }
+            default {
+                return(0x00, returndatasize())
+            }
+        }
+    }
+
+    function _fallback() private {
+        _delegateCall(_getImplementation());
+    }
+
+    fallback() external payable {
+        _fallback();
+    }
+
+    receive() external payable {
+        _fallback();
+    }
+
+    function upgradeTo(address _implementation) external {
+        if (_implementation == _getAdmin()) {
+            revert();
+        }
+
+        _setImplementation(_implementation);
+    }
+
+    function _getAdmin() private view returns (address) {
+        return StorageSlot.getAddressSlot(ADMIN_SLOT).value;
+    }
+
+    function _setAdmin(address _admin) private {
+        if (_admin == address(0)) {
+            revert();
+        }
+
+        StorageSlot.getAddressSlot(ADMIN_SLOT).value = _admin;
+    }
+
+    function _getImplementation() private view returns (address) {
+        return StorageSlot.getAddressSlot(IMPLEMENTATION_SLOT).value;
+    }
+
+    function _setImplementation(address _implementation) private {
+        if (_implementation.code.length == 0) {
+            revert();
+        }
+
+        StorageSlot.getAddressSlot(IMPLEMENTATION_SLOT).value = _implementation;
+    }
+
+    function admin() external view returns (address) {
+        return _getAdmin();
+    }
+
+    function implementation() external view returns (address) {
+        return _getImplementation();
+    }
+}
+
+library StorageSlot {
+    struct AddressSlot {
+        address value;
+    }
+
+    function getAddressSlot(
+        bytes32 _slot
+    ) internal pure returns (AddressSlot storage r) {
+        assembly {
+            r.slot := _slot
+        }
+    }
+}