+++ backend/ignition/modules/RaffleProxy.ts
@@ -1,0 +1,7 @@
+import { buildModule } from '@nomicfoundation/hardhat-ignition/modules';
+
+export default buildModule('RaffleProxy', (m) => {
+  const proxy = m.contract('RaffleProxy', []);
+
+  return { proxy };
+});