+++ backend/ignition/modules/RaffleProxyAdmin.ts
@@ -1,0 +1,7 @@
+import { buildModule } from '@nomicfoundation/hardhat-ignition/modules';
+
+export default buildModule('RaffleProxyAdmin', (m) => {
+  const proxyAdmin = m.contract('RaffleProxyAdmin', []);
+
+  return { proxyAdmin };
+});