+++ client/src/shared/config/contracts.ts
@@ -1,6 +1,6 @@
 import { raffleAbi } from '@/abi/Raffle.abi';
 
 export const raffleContract = {
-  address: '0xE429b5e972D1924eEa4Af628F6c9477F7b0A6F2f',
+  address: '0x22a5c0eB1B81681F7Aae283dbF00A51B43a914A9',
   abi: raffleAbi,
 } as const;