+++ backend/docs/release-notes.md
@@ -22,3 +22,13 @@ Raffle Sepolia: `0xEe2B01764645f876efeE4cD3F34534dDCf58bDDE`
 ## v4
 
 Raffle Sepolia: `0xE429b5e972D1924eEa4Af628F6c9477F7b0A6F2f`
+
+## v5
+
+Added Proxy (Transparent Upgradeable Proxy) for Raffle Contract and Proxy Admin.
+
+Sepolia:
+
+- Raffle `0xE429b5e972D1924eEa4Af628F6c9477F7b0A6F2f`
+- Proxy `0x22a5c0eB1B81681F7Aae283dbF00A51B43a914A9`
+- Proxy Admin `0xeDEF85Bd8E4d073f446E437BcC1209e78753eEe3`